fastapi_graphql_redis/requirements.txt
pandas
fastapi
uvicorn
strawberry-graphql
sqlalchemy
psycopg2-binary
python-dotenv
redis